+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v0.1.2
+
+* Add GUI helpers:
+  * `gui_tab_views()`
+  * `gui_tab_item()`
+* Add `convert_pixel()` to translate `ImageIO` data into B+ terms
+* Fix bug when using `View` instead of `Texture` in CImGui/GuiService
+
 # v0.1.1
 
 * Made OpenGL extension loading more flexible.